fix : 가게평점 소숫점 누락 (int -> float 으로 조회되도록 변경)

+++ Spring/server/src/main/java/spring/reborn/domain/jjim/JjimDao.java
@@ -151,7 +151,7 @@ public class JjimDao {
                         rs.getString("storeName"),
                         rs.getString("storeImage"),
                         rs.getString("category"),
-                        rs.getInt("storeScore")),
+                        rs.getFloat("storeScore")),
                 getJjimStoreListParams
         );
 
@@ -178,7 +178,7 @@ public class JjimDao {
                         rs.getString("storeName"),
                         rs.getString("storeImage"),
                         rs.getString("category"),
-                        rs.getInt("storeScore")),
+                        rs.getFloat("storeScore")),
                 getJjimStoreListParams
         );
